requirements.txt
Pillow==7.0.0
tweepy==3.8.0

